# shared/api/openapi.yaml
openapi: 3.1.0
info:
  title: backend
  description: ''
  license:
    name: ''
  version: 0.1.0
paths:
...
  /projects:
    get:
      tags: []
      operationId: list_projects
      parameters:
      - name: order
        in: query
        required: true
        schema:
          type: string
          enum:
          - last_seen_at_asc
          - last_seen_at_desc
      - name: cursor_last_seen_at
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: cursor_project_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: limit
        in: query
        required: true
        schema:
          type: integer
          format: int32
          maximum: 32
          minimum: 1
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [id, created_by, name, description, indexed_at, last_seen_at, last_seen_file_id]
                      properties:
                        created_by:
                          type: string
                        description:
                          type: string
                        id:
                          type: string
                        indexed_at:
                          type: string
                          format: date-time
                        last_seen_at:
                          type: string
                          format: date-time
                        last_seen_file_id:
                          type: string
                        name:
                          type: string
                  next_cursor:
                    oneOf:
                    - type: 'null'
                    - type: object
                      required: [last_seen_at, project_id]
                      properties:
                        last_seen_at:
                          type: string
                          format: date-time
                        project_id:
                          type: string
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - Bearer: []
    post:
      tags: []
      operationId: create_project
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, description]
              properties:
                description:
                  type: string
                  maxLength: 512
                name:
                  type: string
                  maxLength: 256
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                required: [id, created_by, name, description, indexed_at, last_seen_at, last_seen_file_id]
                properties:
                  created_by:
                    type: string
                  description:
                    type: string
                  id:
                    type: string
                  indexed_at:
                    type: string
                    format: date-time
                  last_seen_at:
                    type: string
                    format: date-time
                  last_seen_file_id:
                    type: string
                  name:
                    type: string
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - Bearer: []
  /projects/{project_id}:
    get:
      tags: []
      operationId: get_project
      parameters:
      - name: project_id
        in: path
        description: Project ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                required: [id, created_by, name, description, indexed_at, last_seen_at, last_seen_file_id]
                properties:
                  created_by:
                    type: string
                  description:
                    type: string
                  id:
                    type: string
                  indexed_at:
                    type: string
                    format: date-time
                  last_seen_at:
                    type: string
                    format: date-time
                  last_seen_file_id:
                    type: string
                  name:
                    type: string
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - Bearer: []
    delete:
      tags: []
      operationId: remove_project
      parameters:
      - name: project_id
        in: path
        description: Project ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Deleted
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - Bearer: []
    patch:
      tags: []
      operationId: update_project
      parameters:
      - name: project_id
        in: path
        description: Project ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type:
                  - string
                  - 'null'
                  maxLength: 512
                name:
                  type:
                  - string
                  - 'null'
                  maxLength: 256
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                required: [id, created_by, name, description, indexed_at, last_seen_at, last_seen_file_id]
                properties:
                  created_by:
                    type: string
                  description:
                    type: string
                  id:
                    type: string
                  indexed_at:
                    type: string
                    format: date-time
                  last_seen_at:
                    type: string
                    format: date-time
                  last_seen_file_id:
                    type: string
                  name:
                    type: string
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - Bearer: []
...
components:
  schemas:
    ErrorBody:
      type: object
      required: [code, description]
      properties:
        code:
          type: string
        description:
          type: string